test(ai): split diplomacy peace tests for repo file-size gate
Refs #2509
