+++ packages/colonizethis_ai/test/diplomacy_planner_stalled_peace_test.dart
@@ -3,9 +3,6 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-import 'domain_planner_test_fake_api.dart';
-import 'planner_test_helpers.dart';
-
 void main() {
   test(
     'stalledExpansionDistractionPeaceTargets peace tribes not focus minor',
@@ -578,449 +575,6 @@ void main() {
           snapshot: aboveStalledThreshold,
         ),
         isEmpty,
-      );
-    },
-  );
-
-  test(
-    'stalledBelowQuotaGpLeadPeaceTargets peace stronger GP while below quota',
-    () {
-      final game = Game(
-        id: 'g-below-quota-gp-lead',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 70),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 8; i++)
-                Province(
-                  id: 'oldWorld|gp3_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp3',
-                ),
-              for (var i = 1; i <= 12; i++)
-                Province(
-                  id: 'oldWorld|gp4_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp4',
-                ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp3', displayName: 'P3', isHuman: false),
-          Player(id: 'gp4', displayName: 'P4', isHuman: false),
-        ],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'gp4',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp3',
-        threats: ThreatSummary(atWarWith: ['gp4', 'minor1']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 8,
-          invadableProvinceIdsSorted: ['oldWorld|inv1'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        stalledBelowQuotaGpLeadPeaceTargets(game: game, snapshot: snapshot),
-        ['gp4'],
-      );
-    },
-  );
-
-  test(
-    'criticalOwHoldPeaceTargets peace all GP wars below quota without minors',
-    () {
-      final game = Game(
-        id: 'g-critical-below-quota-no-minors',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 90),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 6; i++)
-                Province(
-                  id: 'oldWorld|gp3_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp3',
-                ),
-              for (var i = 1; i <= 12; i++)
-                Province(
-                  id: 'oldWorld|gp4_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp4',
-                ),
-              for (var i = 1; i <= 10; i++)
-                Province(
-                  id: 'oldWorld|gp5_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp5',
-                ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp3', displayName: 'P3', isHuman: false),
-          Player(id: 'gp4', displayName: 'P4', isHuman: false),
-          Player(id: 'gp5', displayName: 'P5', isHuman: false),
-        ],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'gp4',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'gp5',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp3',
-        threats: ThreatSummary(atWarWith: ['gp4', 'gp5']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 6,
-          invadableProvinceIdsSorted: ['oldWorld|inv1'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        criticalOwHoldPeaceTargets(game: game, snapshot: snapshot),
-        ['gp4', 'gp5'],
-      );
-    },
-  );
-
-  test(
-    'criticalOwHoldPeaceTargets empty at 7 OW below quota when minors gone',
-    () {
-      final game = Game(
-        id: 'g-critical-below-quota-seven-ow',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 90),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 7; i++)
-                Province(
-                  id: 'oldWorld|gp3_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp3',
-                ),
-              for (var i = 1; i <= 12; i++)
-                Province(
-                  id: 'oldWorld|gp4_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp4',
-                ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp3', displayName: 'P3', isHuman: false),
-          Player(id: 'gp4', displayName: 'P4', isHuman: false),
-        ],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'gp4',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp3',
-        threats: ThreatSummary(atWarWith: ['gp4']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 7,
-          invadableProvinceIdsSorted: ['oldWorld|inv1'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        criticalOwHoldPeaceTargets(game: game, snapshot: snapshot),
-        isEmpty,
-      );
-    },
-  );
-
-  test(
-    'stalledZeroRegimentGpPeaceTargets includes all GP wars when stalled',
-    () {
-      final game = Game(
-        id: 'g-zero-reg-gp-peace',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 80),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 8; i++)
-                Province(
-                  id: 'oldWorld|gp3_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp3',
-                ),
-              for (var i = 1; i <= 12; i++)
-                Province(
-                  id: 'oldWorld|gp4_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp4',
-                ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp3', displayName: 'P3', isHuman: false),
-          Player(id: 'gp4', displayName: 'P4', isHuman: false),
-        ],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'gp4',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp3',
-        threats: ThreatSummary(atWarWith: ['gp4', 'minor1']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 8,
-          invadableProvinceIdsSorted: ['oldWorld|inv1'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        stalledZeroRegimentGpPeaceTargets(game: game, snapshot: snapshot),
-        ['gp4'],
-      );
-    },
-  );
-
-  test(
-    'weakHoldingsInvadableBlockerPeaceTargets peace frontier GP when outmatched',
-    () {
-      final game = Game(
-        id: 'g-weak-blocker-peace',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 60),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 4; i++)
-                Province(
-                  id: 'oldWorld|gp3_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp3',
-                ),
-              for (var i = 1; i <= 11; i++)
-                Province(
-                  id: 'oldWorld|gp4_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp4',
-                ),
-              const Province(
-                id: 'oldWorld|inv1',
-                regionId: 'oldWorld',
-                ownerId: 'gp4',
-              ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp3', displayName: 'P3', isHuman: false),
-          Player(id: 'gp4', displayName: 'P4', isHuman: false),
-        ],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'gp4',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp3',
-        threats: ThreatSummary(atWarWith: ['gp4']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 4,
-          invadableProvinceIdsSorted: ['oldWorld|inv1'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        weakHoldingsInvadableBlockerPeaceTargets(game: game, snapshot: snapshot),
-        ['gp4'],
-      );
-    },
-  );
-
-  test(
-    'criticalWeakUninvadedMinorDeclareTarget picks uninvaded minor owner',
-    () {
-      final game = Game(
-        id: 'g-critical-minor-declare',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 40),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 6; i++)
-                Province(
-                  id: 'oldWorld|gp3_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp3',
-                ),
-              const Province(
-                id: 'oldWorld|m1',
-                regionId: 'oldWorld',
-                ownerId: 'minor1',
-              ),
-              const Province(
-                id: 'oldWorld|m2',
-                regionId: 'oldWorld',
-                ownerId: 'minor2',
-              ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [Player(id: 'gp3', displayName: 'P3', isHuman: false)],
-        minorNations: const [
-          MinorNation(id: 'minor1', displayName: 'M1'),
-          MinorNation(id: 'minor2', displayName: 'M2'),
-        ],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'minor1',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp3',
-        threats: ThreatSummary(atWarWith: ['minor1']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 6,
-          invadableProvinceIdsSorted: ['oldWorld|m1', 'oldWorld|m2'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        criticalWeakUninvadedMinorDeclareTarget(game: game, snapshot: snapshot),
-        'minor2',
-      );
-    },
-  );
-
-  test(
-    'runDiplomacyPlannerWithResult forces peace when candidates are empty',
-    () {
-      final game = Game(
-        id: 'g-empty-candidates-peace',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 60),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 6; i++)
-                Province(
-                  id: 'oldWorld|gp3_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp3',
-                ),
-              for (var i = 1; i <= 10; i++)
-                Province(
-                  id: 'oldWorld|gp4_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp4',
-                ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp3', displayName: 'P3', isHuman: false),
-          Player(id: 'gp4', displayName: 'P4', isHuman: false),
-        ],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'gp4',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-        aiControlByGpId: const {'gp3': true},
-      );
-      const topology = MapTopology(nodes: [], edges: []);
-      final ctx = buildTestPlannerContext(
-        game: game,
-        topology: topology,
-        nationId: 'gp3',
-        primaryGoal: StrategicGoal.trade,
-        suggestionAPI: const FakeOrderSuggestionAPIForDomainPlannerTests(
-          work: [],
-          build: [],
-          move: [],
-          research: [],
-          navalMove: [],
-          navalMission: [],
-          diplomatic: [],
-        ),
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp3',
-        threats: ThreatSummary(atWarWith: ['gp4']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 6,
-          invadableProvinceIdsSorted: ['oldWorld|gp4_10'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      final result = runDiplomacyPlannerWithResult(
-        ctx: ctx,
-        snapshot: snapshot,
-        pass: DiplomacyPlannerPass.nonDeclareWarOnly,
-      );
-      final peaceOrders =
-          result.orders.diplomaticOrdersByPlayerId['gp3'] ?? const [];
-      expect(
-        peaceOrders.any(
-          (o) =>
-              o.type == DiplomaticOrderType.offerPeace &&
-              o.targetFactionId == 'gp4',
-        ),
-        isTrue,
       );
     },
   );
